changed a shitton, readme vv
i got a random sentance chosen, and then chose a random word from that, and matched that word to the part of speech to use for checking. It was much easier to do with two strings rather than trying to work with a 2x2 array so I just changed it.

static and private were causing issues in the classes, I moved some stuff around to make it make sense, you can change back or move things however you want.

i changed the functions to work the way I made it. idk if you want to look at it and try to mimic what I did using your setup, or just go with mine because it's less complicated to work with and change/get words out of.
do whatever you want, just let me know

I have absolutely no fucking idea how to assign random parts of speech to the buttons, while not repeating any, AND while including the correct one. ***we have to create an array to hold the random choices and then assign to the buttons from that array.***

I also made it able to check if the part  they chose is the correct one, but I couldn't make the turns thing work with that idea. Right now they have one chance and then they have to press playAgain.

I just felt like working on this, got a lot done but changed a lot of what you were trying to work with, so let me know