fixed getWord and getWordsPart errors
it now picks a random word correctly and gets its part correctly. For some reason, it is always choosing the last sentence, in sentance[] sentences, but I don't know why

eventually we should make the number of sentences a const that you can just change at the top and the random numbers all adjust?